Add possibility to run cluster generation separatly of full run.
Cluster files are generated in generate methods
Set method can feed the dataHunter directly with clustered data so
it doesnt need to run the cluster generation everytime.

However, for updating the clusters, still need to provide row dataset
and call the generate methods to built the clustering model and
the meta clusters.